Ollama/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

numpy==2.2.6
ollama==0.5.1
openai-whisper==20250625
pyttsx3==2.99
scipy==1.16.0
sounddevice==0.5.2
ugot